Move code from main.py to callback.py.
As it is all @callback code.
